+++ bidreader/__init__.py
@@ -1,3 +1,3 @@
 from .extract import read, Doc
 __all__ = ["read", "Doc"]
-__version__ = "0.8.1"
+__version__ = "0.8.2"